feat: implement resource authorization in tasks endpoint (user can only access and do operations to thier own tasks)

+++ tsilodot-backend-go/service/task.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	"tsilodot/model"
 	"tsilodot/repository"
 )
@@ -12,9 +13,9 @@ type TaskService struct {
 type ITaskService interface {
 	CreateTask(userId uint, task *model.Task) (*model.Task, error)
 	GetTasksByUserID(userId uint, limit int, offset int) ([]model.Task, int64, error)
-	GetTaskByID(taskId uint) (*model.Task, error)
-	UpdateTask(taskId uint, task *model.Task) (*model.Task, error)
-	DeleteTask(taskId uint) error
+	GetTaskByID(taskId uint, userId uint) (*model.Task, error)
+	UpdateTask(taskId uint, userId uint, task *model.Task) (*model.Task, error)
+	DeleteTask(taskId uint, userId uint) error
 }
 
 func NewTaskService(taskRepository repository.ITaskRepository) ITaskService {
@@ -30,14 +31,27 @@ func (s *TaskService) GetTasksByUserID(userId uint, limit int, offset int) ([]mo
 	return s.taskRepository.FindTasksByUserID(nil, userId, limit, offset)
 }
 
-func (s *TaskService) GetTaskByID(taskId uint) (*model.Task, error) {
-	return s.taskRepository.FindTaskByID(nil, taskId)
-}
-
-func (s *TaskService) UpdateTask(taskId uint, taskUpdate *model.Task) (*model.Task, error) {
+func (s *TaskService) GetTaskByID(taskId uint, userId uint) (*model.Task, error) {
 	task, err := s.taskRepository.FindTaskByID(nil, taskId)
 	if err != nil {
 		return nil, err
+	}
+
+	if task.UserID != userId {
+		return nil, errors.New("unauthorized: task does not belong to user")
+	}
+
+	return task, nil
+}
+
+func (s *TaskService) UpdateTask(taskId uint, userId uint, taskUpdate *model.Task) (*model.Task, error) {
+	task, err := s.taskRepository.FindTaskByID(nil, taskId)
+	if err != nil {
+		return nil, err
+	}
+
+	if task.UserID != userId {
+		return nil, errors.New("unauthorized: cannot update task belonging to another user")
 	}
 
 	task.Title = taskUpdate.Title
@@ -48,6 +62,15 @@ func (s *TaskService) UpdateTask(taskId uint, taskUpdate *model.Task) (*model.Ta
 	return s.taskRepository.UpdateTask(nil, task)
 }
 
-func (s *TaskService) DeleteTask(taskId uint) error {
+func (s *TaskService) DeleteTask(taskId uint, userId uint) error {
+	task, err := s.taskRepository.FindTaskByID(nil, taskId)
+	if err != nil {
+		return err
+	}
+
+	if task.UserID != userId {
+		return errors.New("unauthorized: cannot delete task belonging to another user")
+	}
+
 	return s.taskRepository.DeleteTask(nil, taskId)
 }
